pyproject.toml
[tool.poetry]
name = "auth-service"
version = "0.1.0"
description = ""
authors = ["Antonio Luca <toolanto@gmail.com>"]
readme = "README.md"
packages = [{include = "auth_service", from = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.100.0"
pydantic = "^2.0.2"
aiofiles = "^23.1.0"
uvicorn = "^0.22.0"
python-dotenv="^1.0.0"
pydantic-settings = "^2.0.1"
pyjwt = "^2.7.0"


[tool.poetry.group.test.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
pytest-mock = "^3.11.1"
pytest-cov = "^4.1.0"
factory-boy = "^3.2.1"
httpx = "^0.24.1"
pre-commit = "^3.3.3"
black = "^23.3.0"
isort = "^5.11.3"
flake8 = "^6.0.0"


[tool.black]
line-length = 100
target-version = ['py39']

[tool.isort]
profile = "black"
multi_line_output = 3
py_version = 39

[tool.pytest.ini_options]
markers = [
    "integration"
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
